move replace_contents
